net-misc/wdt-9999: fix dependencies and install
wdt uses the deprecated namespace "google" in gflags. To ensure
compatibility with other packages that use gflags, version 2.1.2 or
greater of gflags must be used as it should fully support both old
("google") and new ("gflags") namespaces.

This ebuild for wdt also strips out the folly library pieces and relies
on a shared library. However, the current version of folly does not
build correctly with all components so wdt must utilize a customized
folly library until changes are accepted upstream.